backend/requirements.txt
torch>=2.1.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=0.3.0

